feature: AP-10-Create new Controlled input component as was already using on sign in page, refactored sign in page following new component
