+++ src/components/atoms/index.ts
@@ -1,1 +1,2 @@
 export { APFab } from "./APFab/APFab";
+export { APControlledInput } from "./APControlledInput/APControlledInput";
