+++ src/components/organisms/NavigationMenu/index.ts
@@ -1,0 +1,1 @@
+export { NAVIGATION_PATH } from "./constants/navigationPath";
