+++ src/pages/SignInPage/SignInPage.tsx
@@ -1,19 +1,18 @@
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Container from "@mui/material/Container";
-import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 import Stack from "@mui/material/Stack";
 import { useNavigate } from "react-router-dom";
-import { Controller, useForm } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import { schema } from "./login.schema";
-import { StyledSignInPaper } from "./StyledSignInPage.styled";
 import { yupResolver } from "@hookform/resolvers/yup";
 import type { IFormData } from "./types";
 import { useDispatch } from "react-redux";
-import { useLoginUserMutation } from "../../services/auth/authApi";
-import { logout, setCredentials } from "../../services/auth/authState";
-import { NAVIGATION_PATH } from "@/components/organisms/NavigationMenu/constants/navigationPath";
+import { APControlledInput } from "@/components/atoms";
+import { NAVIGATION_PATH } from "@/components/organisms/NavigationMenu";
+import { logout, setCredentials, useLoginUserMutation } from "@/services/auth";
+import { StyledSignInPaper } from "./StyledSignInPage.styled";
 
 export const SignInPage = () => {
   const [loginUser, { isLoading }] = useLoginUserMutation();
@@ -50,31 +49,21 @@ export const SignInPage = () => {
           <Box component="form" onSubmit={handleSubmit(onSubmit)} noValidate>
             <Stack spacing={2}>
               <Typography variant="h5">Sign in</Typography>
-              <Controller
+              <APControlledInput
                 control={control}
                 name="username"
-                render={({ field }) => (
-                  <TextField
-                    {...field}
-                    placeholder="Username"
-                    type="email"
-                    error={!!errors.username}
-                    helperText={errors.username?.message}
-                  />
-                )}
+                placeholder="Username"
+                type="email"
+                error={!!errors.username}
+                helperText={errors.username?.message}
               />
-              <Controller
+              <APControlledInput
                 control={control}
                 name="password"
-                render={({ field }) => (
-                  <TextField
-                    {...field}
-                    placeholder="Password"
-                    type="password"
-                    error={!!errors.password}
-                    helperText={errors.password?.message}
-                  />
-                )}
+                placeholder="Password"
+                type="password"
+                error={!!errors.password}
+                helperText={errors.password?.message}
               />
               <Button type="submit" variant="contained" disabled={isLoading}>
                 {isLoading ? "Loading..." : "Submit"}
