+++ src/services/auth/index.ts
@@ -1,4 +1,7 @@
-export { logout } from "./authState";
-export { selectAccessToken } from "./constants/authSelectors";
-export { selectRefreshToken } from "./constants/authSelectors";
-export { selectAuthUser } from "./constants/authSelectors";
+export { logout, setCredentials } from "./authState";
+export {
+  selectAccessToken,
+  selectRefreshToken,
+  selectAuthUser,
+} from "./constants/authSelectors";
+export { useLoginUserMutation } from "./authApi";
